requirements.txt
Pillow
configparser